Return to stable builds
This patch removes a line in our config where we purposefully forget about previous builds to ensure a clean from scratch deploy. This was to facilitate our improved collateral system and cleaned up migrations. Once that deploy is finished and stable, we should pull this PR to return to incremental deployments.